+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 
 ### unreleased
+- **doc:** add documentation
 
 ### [v0.4.0] - 2016-07-16
 - **fix:** version in meta data