+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### unreleased
 - **doc:** add documentation
 - **doc:** improve plugin options related paragraphs
+- **fix:** linting error
 
 ### [v0.4.0] - 2016-07-16
 - **fix:** version in meta data