+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **fix:** adjust tests
 - **doc:** adjust documentation
 - **fix:** make links halicious
+- **fix:** adjust tests
 
 ### [v0.4.0] - 2016-07-16
 - **fix:** version in meta data